+++ api-agendamento-serviços/src/main/java/com/automotiva/estetica/rick/application/controller/OrdemServicoController.java
@@ -2,14 +2,11 @@ package com.automotiva.estetica.rick.application.controller;
 
 import com.automotiva.estetica.rick.application.dto.request.OrdemServicoRequest;
 import com.automotiva.estetica.rick.application.dto.request.PageRequest;
-import com.automotiva.estetica.rick.application.dto.response.AgendamentosHojeListResponse;
 import com.automotiva.estetica.rick.application.dto.response.HorarioDisponivelResponse;
 import com.automotiva.estetica.rick.application.dto.response.OrdemServicoResponse;
-import com.automotiva.estetica.rick.application.mapper.AgendamentoHojeMapper;
 import com.automotiva.estetica.rick.application.security.ClienteOnly;
 import com.automotiva.estetica.rick.application.security.OwnershipValidator;
 import com.automotiva.estetica.rick.application.service.OrdemServicoApplicationService;
-import com.automotiva.estetica.rick.domain.gateway.ItemServicoGateway;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -30,8 +27,6 @@ public class OrdemServicoController {
 
     private final OrdemServicoApplicationService ordemServicoUseCase;
     private final OwnershipValidator ownershipValidator;
-    private final ItemServicoGateway itemServicoGateway;
-    private final AgendamentoHojeMapper agendamentoHojeMapper;
 
     @GetMapping
     @Operation(summary = "Lista todas as ordens de serviço paginadas")
@@ -73,9 +68,4 @@ public class OrdemServicoController {
         return ResponseEntity.ok(ordemServicoUseCase.buscarHorariosDisponiveis(data, servicosIds));
     }
 
-    @GetMapping("/hoje")
-    @Operation(summary = "Lista todos os agendamentos do dia atual com detalhes completos para visualização no dashboard")
-    public ResponseEntity<AgendamentosHojeListResponse> buscarAgendamentosHoje() {
-        return ResponseEntity.ok(ordemServicoUseCase.buscarAgendamentosHoje());
-    }
 }
